split tests into logical modules
